# docs/api-specs-process-2.yaml
openapi: 3.0.0
info:
  title: Process 2 - Group & Advisor Formation API
  version: 1.1.0
  description: >
    This API manages the processes of students forming groups, the Team Leader setting up 
    GitHub/JIRA integrations, and sending advisee requests to professors. It also allows 
    system coordinators (admins) to intervene and manage groups manually.
servers:
  - url: http://localhost:3000
    description: Local Development Server

paths:
  /api/groups:
    post:
      summary: 2.0 Create a New Group
      description: >
        Triggered by a student logging into the system. The system first checks if the student 
        is valid in the D1: Users datastore. Once the group is successfully created, the requesting 
        student is automatically assigned as the 'Team Leader', and the data is saved to the 
        D2: Groups datastore.
      tags:
        - Group Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                groupName:
                  type: string
            example:
              studentId: "std-12345"
              groupName: "Tech Titans"
      responses:
        "201":
          description: Group created successfully.
          content:
            application/json:
              example:
                message: "Group created successfully"
                groupId: "grp-98765"
                teamLeader: "std-12345"
        "400":
          description: Missing or invalid data provided.
          content:
            application/json:
              example:
                error: "Bad Request: 'groupName' and 'studentId' fields are strictly required."
        "409":
          description: Conflict. Group name is taken or student is already in a group.
          content:
            application/json:
              example:
                error: "Conflict: The group name 'Tech Titans' is already taken, or the student is already assigned to another group."

  /api/groups/{groupId}/members:
    put:
      summary: 2.0 Add or Remove Group Member
      description: >
        Allows the Team Leader to manage their group. The student ID and action type (add/remove) 
        are required. The system checks if the student to be added is already a member of another group.
      tags:
        - Group Management
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                action:
                  type: string
                  enum: [add, remove]
            example:
              studentId: "std-54321"
              action: "add"
      responses:
        "200":
          description: Member successfully added or removed.
          content:
            application/json:
              example:
                message: "Student std-54321 successfully added to the group."
                currentMembers: ["std-12345", "std-54321"]
        "403":
          description: Unauthorized action.
          content:
            application/json:
              example:
                error: "Forbidden: Only the Team Leader of this group can add or remove members."
        "404":
          description: Student not found.
          content:
            application/json:
              example:
                error: "Not Found: Student ID 'std-54321' does not exist in the system."

  /api/groups/{groupId}/integrations:
    post:
      summary: 2.2 Setup GitHub and JIRA Integration
      description: >
        The Team Leader saves the GitHub Personal Access Token (PAT) and JIRA Workspace URL 
        required for the daily AI Sync & Audit processes. This information is securely stored 
        in the D8: Integration Tokens datastore.
      tags:
        - Integrations
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                githubPat:
                  type: string
                jiraSpaceUrl:
                  type: string
            example:
              githubPat: "ghp_abc123def456ghi789..."
              jiraSpaceUrl: "https://myuni.atlassian.net/wiki/spaces/TECH"
      responses:
        "201":
          description: Tokens saved successfully.
          content:
            application/json:
              example:
                message: "Integration tokens securely saved to D8 Datastore."
        "400":
          description: Invalid token or URL format.
          content:
            application/json:
              example:
                error: "Bad Request: Invalid GitHub PAT format or JIRA URL is unreachable."
        "403":
          description: Unauthorized access.
          content:
            application/json:
              example:
                error: "Forbidden: You must be the Team Leader to set up tool integrations."

  /api/groups/{groupId}/advisor-requests:
    post:
      summary: 2.3 Send Advisee Request to Professor
      description: >
        Once the group formation is complete, the Team Leader sends an advisee request to a 
        professor selected from the list. If a request is currently pending, the group is blocked 
        from sending requests to other professors.
      tags:
        - Advisor Association
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                professorId:
                  type: string
            example:
              professorId: "prof-404"
      responses:
        "201":
          description: Request successfully sent.
          content:
            application/json:
              example:
                message: "Advisor request sent successfully to Professor prof-404."
                requestId: "req-888"
        "404":
          description: Professor not found.
          content:
            application/json:
              example:
                error: "Not Found: No professor found with ID 'prof-404'."
        "409":
          description: Group already in the advisor assignment process.
          content:
            application/json:
              example:
                error: "Conflict: This group already has a pending advisor request or is already assigned to an advisor."

  /api/groups/{groupId}/advisor-requests/{requestId}:
    put:
      summary: 2.3 Approve/Decline Advisor Request
      description: >
        Can only be used by the target Professor. The professor reviews the incoming request and 
        makes a decision. If approved, the group is linked to this professor in the D2 datastore.
      tags:
        - Advisor Association
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision:
                  type: string
                  enum: [approve, decline]
            example:
              decision: "approve"
      responses:
        "200":
          description: Decision processed successfully.
          content:
            application/json:
              example:
                message: "Group is now successfully assigned to you."
                status: "approved"
        "403":
          description: Unauthorized action (Wrong professor).
          content:
            application/json:
              example:
                error: "Forbidden: You cannot respond to a request directed at another professor."

  /api/admin/groups/{groupId}/transfer:
    put:
      summary: 2.3 (Admin) Transfer Group to Another Advisor
      description: >
        The system coordinator (Admin) can manually transfer a group to another professor in 
        exceptional situations (e.g., professor leaving). This action bypasses the standard 
        approval mechanisms.
      tags:
        - Advisor Association (Admin)
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newProfessorId:
                  type: string
            example:
              newProfessorId: "prof-505"
      responses:
        "200":
          description: Transfer successful.
          content:
            application/json:
              example:
                message: "Group has been manually transferred to new advisor."
                newAdvisorId: "prof-505"
        "401":
          description: Missing admin privileges.
          content:
            application/json:
              example:
                error: "Unauthorized: Only system coordinators/admins can perform manual transfers."

  /api/projects:
    get:
      summary: Lookup the projects (course templates) available to students
      description: >
        This endpoint enables students to look up the available general projects (templates).
      tags:
        - General Projects
        - Templates
      responses:
        "200":
          description: >
            Lookup successful. General projects are included in response.
          content:
            application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  projectId:
                    type: string
                  name:
                    type: string
                  sprintCount:
                    type: number
                  deliverables:
                    type: string
            example:            
              - projectId: "GP - 001"
                name: "Senioritis Project"
                sprintCount: 5
                deliverables: >
                  "Students will deliver the most amazing project that wows all observers"
              - projectId: "GP - 002"
                name: "Senioritis Projectile"
                sprintCount: 3
                deliverables: >
                  "Students will deliver the most amazing project that wows all observers really, really fast"
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                error: "The server was unable to process your request. Please contact the server administrator."